+++ tests/Header/Parameter/ParameterTest.php
@@ -10,20 +10,20 @@ class ParameterTest extends TestCase
 {
     public function testInterface()
     {
-        $p = new Parameter('q', 'foo');
+        $p = Parameter::of('q', 'foo');
 
         $this->assertSame('q', $p->name());
         $this->assertSame('foo', $p->value());
         $this->assertSame('q=foo', $p->toString());
 
-        $this->assertSame('level', (new Parameter('level', ''))->toString());
+        $this->assertSame('level', (Parameter::of('level', ''))->toString());
     }
 
     public function testQuoteWhenThereIsAWithespace()
     {
         $this->assertSame(
             'foo="bar baz"',
-            (new Parameter('foo', 'bar baz'))->toString(),
+            (Parameter::of('foo', 'bar baz'))->toString(),
         );
     }
 
@@ -31,7 +31,7 @@ class ParameterTest extends TestCase
     {
         $this->assertSame(
             "foo=\"bar\tbaz\"",
-            (new Parameter('foo', "bar\tbaz"))->toString(),
+            (Parameter::of('foo', "bar\tbaz"))->toString(),
         );
     }
 
@@ -39,7 +39,7 @@ class ParameterTest extends TestCase
     {
         $this->assertSame(
             'foo="bar baz"',
-            (new Parameter('foo', '"bar baz"'))->toString(),
+            (Parameter::of('foo', '"bar baz"'))->toString(),
         );
     }
 
@@ -47,7 +47,7 @@ class ParameterTest extends TestCase
     {
         $this->assertSame(
             'foo="bar"',
-            (new Parameter('foo', '"bar"'))->toString(),
+            (Parameter::of('foo', '"bar"'))->toString(),
         );
     }
 }